% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ops_wrapper.R
\name{agg_count}
\alias{agg_count}
\title{Number of observations by group}
\arguments{
\item{col}{String denoting the name of the new column to hold the counts of each aggregation group.}
}
\value{
\code{AggOp} to be used in a call to \code{agg_by()}.
}
\description{
Creates a Count aggregation that counts the number of rows in each aggregation group.
}
\details{
The aggregation groups that this function acts on are defined with the \code{by} parameter of the \code{agg_by()} caller
function. The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the \code{A} and
\code{B} columns.

This function, like other Deephaven \code{agg} functions, is a generator function. That is, its output is another
function called an \code{AggOp} intended to be used in a call to \code{agg_by()} or \code{agg_all_by()}. This detail is
typically hidden from the user by the \code{agg_by()} and \code{agg_all_by()} functions, which call the \code{AggOp} internally.
However, it is important to understand this detail for debugging purposes, as the output of an \code{agg} function can
otherwise seem unexpected.

Note that this operation is not supported in \code{agg_all_by()}.
}
\examples{
print("hello!")

}
