% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ops_wrapper.R
\name{agg_std}
\alias{agg_std}
\title{Standard deviation of specified columns by group}
\arguments{
\item{cols}{String or list of strings denoting the column(s) to aggregate. Can be renaming expressions, i.e. “new_col = col”.
Default is to aggregate all non-grouping columns, which is only valid in the \code{agg_all_by()} operation.}
}
\value{
Aggregation function to be used in \code{agg_by()} or \code{agg_all_by()}.
}
\description{
Creates a Standard Deviation aggregation that computes the standard deviation of each column in \code{cols}, for each aggregation group.
}
\details{
The aggregation groups that this function acts on are defined with the \code{by} parameter of the \code{agg_by()} or
\code{agg_all_by()} caller function. The aggregation groups are defined by the unique combinations of values in the \code{by}
columns. For example, if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of
values in the \code{A} and \code{B} columns.

This function, like the other Deephaven \verb{agg_*} functions, is a generator function. That is, its output is another
function that is intended to be used in a call to \code{agg_by()} or \code{agg_all_by()}. This detail is typically hidden from
the user by the \code{agg_by()} and \code{agg_all_by()} functions, which call the generated functions internally. However, it is
important to understand this detail for debugging purposes, as the output of an \verb{agg_*} function can otherwise seem
unexpected.
}
\examples{
print("hello!")

}
