% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ops_wrapper.R
\docType{class}
\name{AggBy}
\alias{AggBy}
\title{Aggregations in Deephaven}
\description{
Table aggregations are a quintessential feature of Deephaven. You can apply as many aggregations as
needed to static tables \emph{or} streaming tables, and if the parent tables are streaming, the resulting aggregated
tables will update alongside their parent tables. It is also very easy to perform \emph{grouped} aggregations, which
allow you to aggregate tables on a per-group basis.
}
\section{Apply aggregations to a table}{

There are two methods for performing aggregations on a table, \code{agg_by()} and \code{agg_all_by()}. \code{agg_by()} allows you to
perform many aggregations on specified columns, while \code{agg_all_by()} allows you to perform a single aggregation to
every column in the table. Both methods have an optional \code{by} parameter that is used to specify grouping columns.
Here are some details on each method:
\itemize{
\item \code{TableHandle$agg_by(aggs, by)}: Creates a new table containing grouping columns and grouped data.
The resulting grouped data is defined by the aggregation(s) specified.
\item \code{TableHandle$agg_all_by(agg, by)}: Creates a new table containing grouping columns and grouped data.
The resulting grouped data is defined by the aggregation specified. This method applies the aggregation to all
columns of the table, so it can only accept one aggregation at a time.
}

The \code{agg_by()} and \code{agg_all_by()} methods themselves do not know anything about the columns you want to perform
aggregations on. Rather, the desired columns are passed to individual \code{agg} functions, enabling you to apply
various kinds of aggregations to different columns or groups of columns as needed.
}

\section{\code{agg} functions}{

\code{agg} functions are used to perform aggregation calculations on grouped data by passing them to \code{agg_by()} or
\code{agg_all_by()}. These functions are \emph{generators}, meaning they return \emph{functions} that the Deephaven engine knows
how to interpret. We call the functions that they return \code{AggOp}s. These \code{AggOp}s are not R-level functions, but
wrappers around C++ functions that perform all of the intensive calculations. Here is a list of all \code{agg} functions
available in Deephaven:
\itemize{
\item \code{\link[=agg_first]{agg_first()}}
\item \code{\link[=agg_last]{agg_last()}}
\item \code{\link[=agg_min]{agg_min()}}
\item \code{\link[=agg_max]{agg_max()}}
\item \code{\link[=agg_sum]{agg_sum()}}
\item \code{\link[=agg_abs_sum]{agg_abs_sum()}}
\item \code{\link[=agg_avg]{agg_avg()}}
\item \code{\link[=agg_w_avg]{agg_w_avg()}}
\item \code{\link[=agg_median]{agg_median()}}
\item \code{\link[=agg_var]{agg_var()}}
\item \code{\link[=agg_std]{agg_std()}}
\item \code{\link[=agg_percentile]{agg_percentile()}}
\item \code{\link[=agg_count]{agg_count()}}
}

For more details on each aggregation function, click on one of the methods above or see the reference documentation
by running \code{?agg_first}, \code{?agg_last}, etc.
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{NULL
}\if{html}{\out{</div>}}
}

