% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_wrapper.R
\docType{class}
\name{Aggregation}
\alias{Aggregation}
\title{Deephaven Aggregations}
\arguments{
\item{cols}{String or list of strings denoting the column(s) to aggregate. Can be renaming expressions, i.e. “new_col = col”.
Default is to aggregate all non-grouping columns, which is only valid in the `agg_all_by()` operation.}
}
\value{
Aggregation function to be used in `agg_by()` or `agg_all_by()`.
}
\description{
An Aggregation represents an aggregation operation that can be passed to `agg_by()` or `agg_all_by()`.
Note that Aggregations should not be instantiated directly by user code, but rather by provided agg_* functions.

Creates a First aggregation that computes the first value of each column in `cols` for each aggregation group.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Aggregation-new}{\code{Aggregation$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Aggregation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Aggregation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Aggregation$new(aggregation, agg_name, ...)}\if{html}{\out{</div>}}
}

}
}
