% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_ops_wrapper.R
\docType{class}
\name{Aggregations}
\alias{Aggregations}
\title{Aggregations}
\description{
Table aggregations are a quintessential feature of Deephaven. You can apply as many aggregations as
needed to static tables \emph{or} streaming tables, and if the parent tables are streaming, the resulting aggregated
tables will update alongside their parent tables. It is also very easy to perform \emph{grouped} aggregations, which
allow you to aggregate tables on a per-group basis.
}
\section{Apply aggregations to a table}{

There are two methods for performing aggregations on a table, \code{agg_by()} and \code{agg_all_by()}. \code{agg_by()} allows you to
perform many aggregations on specified columns, while \code{agg_all_by()} allows you to perform a single aggregation to
every column in the table. Both methods have an optional \code{by} parameter that is used to specify grouping columns.
Here are some details on each method:
\itemize{
\item \code{TableHandle$agg_by(aggs, by)}: Creates a new table containing grouping columns and grouped data.
The resulting grouped data is defined by the aggregation(s) specified.
\item \code{TableHandle$agg_all_by(agg, by)}: Creates a new table containing grouping columns and grouped data.
The resulting grouped data is defined by the aggregation specified. This method applies the aggregation to all
columns of the table, so it can only accept one aggregation at a time.
}

For more details, see the reference documentation on each method by running \code{?agg_by}, or \code{?agg_all_by}.
}

\section{Aggregation functions}{

Aggregation functions are used to define the grouped data in an aggregated table by passing them to \code{agg_by()}
or \code{agg_all_by()}. These functions are \emph{generators}, meaning they return \emph{functions} that the Deephaven engine knows
how to interpret. Here is a list of all aggregation functions available in Deephaven:
\itemize{
\item \code{\link[=agg_first]{agg_first(cols)}}: Creates a First aggregation that computes the first value of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_last]{agg_last(cols)}}: Creates a Last aggregation that computes the last value of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_min]{agg_min(cols)}}: Creates a Minimum aggregation that computes the minimum of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_max]{agg_max(cols)}}: Creates a Maximum aggregation that computes the maximum of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_sum]{agg_sum(cols)}}: Creates a Sum aggregation that computes the sum of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_abs_sum]{agg_abs_sum(cols)}}: Creates an Absolute Sum aggregation that computes the absolute sum of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_avg]{agg_avg(cols)}}: Creates an Average aggregation that computes the average of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_w_avg]{agg_w_avg(wcol, cols)}}: Creates a Weighted Average aggregation that computes the weighted average of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_median]{agg_median(cols)}}: Creates a Median aggregation that computes the median of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_var]{agg_var(cols)}}: Creates a Variance aggregation that computes the variance of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_std]{agg_std(cols)}}: Creates a Standard Deviation aggregation that computes the standard deviation of each column in \code{cols}, for each aggregation group.
\item \code{\link[=agg_percentile]{agg_percentile(percentile, cols)}}: Creates a Percentile aggregation that computes the given percentile of each column in \code{cols} for each aggregation group.
\item \code{\link[=agg_count]{agg_count(col)}}: Creates a Count aggregation that computes the number of rows in each aggregation group.
}

For more details on each aggregation function, click on one of the methods above or see the reference documentation
by running \code{?agg_first}, \code{?agg_last}, etc.
}

\section{Usage}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{NULL
}\if{html}{\out{</div>}}
}

