% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{arrow_to_dh_table}
\alias{arrow_to_dh_table}
\title{Push an Arrow Table to the server as a Deephaven Table.}
\usage{
arrow_to_dh_table(client, arrow_tbl, name)
}
\arguments{
\item{client}{A Deephaven Client object maintaining the connection to the server.}

\item{name}{A string containing the name of the new Deephaven Table to be created on the server.}

\item{rbr}{An Arrow Table containing the data to be converted to a Deephaven Table.}
}
\value{
A Deephaven TableHandle, providing a reference to the new table on the server.
}
\description{
Converts an Arrow Table to a Deephaven Table, pushes that new table to the Deephaven server,
and returns a TableHandle corresponding to the new table.
}
