% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dh_to_arrow_table}
\alias{dh_to_arrow_table}
\title{Convert a Deephaven TableHandle to an Arrow Table.}
\usage{
dh_to_arrow_table(tableHandle)
}
\arguments{
\item{tableHandle}{A Deephaven TableHandle.}
}
\value{
An Arrow Table.
}
\description{
Converts a Deephaven TableHandle to an [Arrow Table](https://arrow.apache.org/docs/r/reference/Table.html),
R data frames.
}
