% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dh_to_data_frame}
\alias{dh_to_data_frame}
\title{Convert a Deephaven TableHandle to an R Data Frame.}
\usage{
dh_to_data_frame(tableHandle)
}
\arguments{
\item{tableHandle}{A Deephaven TableHandle.}
}
\value{
An R Data Frame.
}
\description{
Converts a Deephaven TableHandle to an R Data Frame.
This is the recommended method for converting a Deephaven TableHandle to something that is usable in R.
}
