% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dh_to_rbs_reader}
\alias{dh_to_rbs_reader}
\title{Convert a Deephaven TableHandle to an Arrow RecordBatchStreamReader}
\usage{
dh_to_rbs_reader(tableHandle)
}
\arguments{
\item{tableHandle}{A Deephaven TableHandle.}
}
\value{
An Arrow RecordBatchStreamReader.
}
\description{
Converts a Deephaven TableHandle to an [Arrow RecordBatchStreamReader](https://arrow.apache.org/docs/r/reference/RecordBatchReader.html),
which can itself be used to produce Arrow tables and R data frames.
}
