% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_handle_wrapper.R
\name{merge_tables}
\alias{merge_tables}
\title{Merge tables with the same schema}
\usage{
merge_tables(...)
}
\arguments{
\item{...}{Arbitrary number of TableHandles or vectors of TableHandles with a schema matching this table.}
}
\value{
A TableHandle referencing the new table.
}
\description{
Merges several tables into one table on the server. All tables must have the same schema, and can
be supplied as a list of TableHandles, any number of TableHandles, or a mix of both.
}
\examples{
print("hello!")

}
