% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_control.R
\name{op_control}
\alias{op_control}
\title{Handling special values in some \code{update_by()} operations}
\usage{
op_control(on_null = "skip", on_nan = "skip", big_value_context = "decimal128")
}
\arguments{
\item{on_null}{Defines how an UpdateByOp handles null values it encounters. \code{'skip'} is the default.}

\item{on_nan}{Defines how an UpdateByOp handles NaN values it encounters. \code{'skip'} is the default.}

\item{big_value_context}{Defines how an UpdateByOp handles exceptionally large values it encounters.
\code{'decimal128'} is the default.}
}
\value{
OperationControl to be used in \verb{uby_em*}.
}
\description{
Creates an OperationControl that controls behavior on special values in \verb{uby_em*} operations.
}
\details{
An OperationControl defines the control parameters of some UpdateByOps used in an update_by table operation.
The UpdateBy operations that can use OperationControl to handle erroneous data are:
\itemize{
\item \code{\link[=ema_tick]{ema_tick()}}
\item \code{\link[=ema_time]{ema_time()}}
\item \code{\link[=ems_tick]{ems_tick()}}
\item \code{\link[=ems_time]{ems_time()}}
\item \code{\link[=emmin_tick]{emmin_tick()}}
\item \code{\link[=emmin_time]{emmin_time()}}
\item \code{\link[=emmax_tick]{emmax_tick()}}
\item \code{\link[=emmax_time]{emmax_time()}}
}

The arguments \code{on_null} and \code{on_nan} can take the following values:
\itemize{
\item \code{'poison'}: Allow bad data to poison the result. This is only valid for use with NaN.
\item \code{'reset'}: Reset the state for the bucket to NULL when invalid data is encountered.
\item \code{'skip'}: Skip and do not process the invalid data without changing state.
\item \code{'throw'}: Throw an exception and abort processing when bad data is encountered.
}

The argument \code{big_value_context} can take the following values:
\itemize{
\item \code{'decimal128'}: IEEE 754R Decimal128 format. 34 digits and rounding is half-even.
\item \code{'decimal32'}: IEEE 754R Decimal32 format. 7 digits and rounding is half-even.
\item \code{'decimal64'}: IEEE 754R Decimal64 format. 16 digits and rounding is half-even.
\item \code{'unlimited'}: Unlimited precision arithmetic. Rounding is half-up.
}

This function is a generator function. That is, its output is another function that is intended to be used in a call
to one of the above \code{update_by()} methods. This detail is typically hidden from the user by \code{update_by()}, which calls
the generated functions internally. However, it is important to understand this detail for debugging purposes, as the
output of a call to \code{op_control()} can otherwise seem unexpected.
}
\examples{
print("hello!")

}
