% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_control.R
\docType{class}
\name{OperationControl}
\alias{OperationControl}
\title{Deephaven OperationControl}
\description{
An \code{OperationControl} is the return type of Deephaven's \code{\link{op_control}} function. It is a function that
determines how special values will be handled in the context of an \code{\link[=UpdateBy]{update_by}} operation.
An \code{OperationControl} is intended to be passed directly to a \code{uby} function, and should never be instantiated
directly be user code.

If you plan to use the same operation control parameters for multiple \code{uby} functions in one or more \code{update_by} calls,
consider creating an \code{OperationControl} object and reusing it. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{opc <- op_control(on_null = 'skip', on_nan = 'poison', big_value_context = 'decimal64')

result <- th$update_by(c(uby_ema_tick(5, c("XEma = X", "YEma = Y"), opc),
                         uby_emstd_tick(5, c("XEmstd = X", "YEmstd = Y"), opc)),
                       by="Group")
}\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperationControl-new}{\code{OperationControl$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationControl-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperationControl-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationControl$new(op_control)}\if{html}{\out{</div>}}
}

}
}
