% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rbr_to_dh_table}
\alias{rbr_to_dh_table}
\title{Push an Arrow RecordBatchReader to the server as a Deephaven Table.}
\usage{
rbr_to_dh_table(client, rbr, name)
}
\arguments{
\item{client}{A Deephaven Client object maintaining the connection to the server.}

\item{rbr}{An Arrow RecordBatchReader containing the data to be converted to a Deephaven Table.}

\item{name}{A string containing the name of the new Deephaven Table to be created on the server.}
}
\value{
A Deephaven TableHandle, providing a reference to the new table on the server.
}
\description{
Converts an Arrow RecordBatchReader to a Deephaven Table, pushes that new table to the Deephaven server,
and returns a TableHandle corresponding to the new table.
}
