% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{rdeephaven}
\alias{rdeephaven}
\title{The Deephaven Community R Client}
\description{
The Deephaven Community R Client provides an R interface to Deephaven's powerful real-time data engine, \href{https://deephaven.io/community/}{\emph{Deephaven Core}}.
To use this package, you must have a Deephaven server running and be able to connect to it. For more information on
how to set up a Deephaven server, see the documentation \href{https://deephaven.io/core/docs/tutorials/quickstart/}{here}.
For a brief overview of the Deephaven R Client, see the \href{https://deephaven.io/blog/2023/10/04/r-client-updates/}{blog post}
we published on the topic.
}
\section{Building blocks of the Deephaven R Client}{

There are two primary R classes that make up the Deephaven R Client, the \code{\link{Client}} class and the
\code{\link{TableHandle}} class. The \code{Client} class is used to establish a connection to the Deephaven server with
its constructor \code{Client$new()}, and to send server requests, such as running a script via \code{run_script()}, or pushing
local data to the server via \code{open_table()}. Many of these server requests end up creating or modifying tables that
live on the server. To keep track of these tables, the R client retrieves references to them, and wraps these references
in \code{TableHandle} objects. These TableHandles have a host of methods that mirror server-side table operations, such as
\code{head()}, \code{tail()}, \code{update_by()}, and so on. So, you can typically use TableHandles \emph{as if} they are tables themselves,
and all of the corresponding methods that you call on them will be executed on the server. TableHandles also support
common functional methods for converting server-side Deephaven tables to R objects stored in local memory such as
\code{as.data.frame()}, \code{as_tibble()}, and \code{as_arrow_table()}. For more information on these classes and all of their methods,
see the reference documentation for \code{\link{Client}} and \code{\link{TableHandle}} by clicking on their class names,
or by running \code{?Client} or \code{?TableHandle}.
}

\section{Real-time data analysis}{

Since TableHandles are references to tables living on the Deephaven server, they may refer to streaming tables, or
or tables that are receiving new data every second. R objects like data frames or Dplyr tibbles do not have this property -
they are always static objects stored in memory. However, a TableHandle referring to a streaming table may be converted
to a data frame or tibble at any time, and the resulting object will be a snapshot of the table at the time of conversion.
This means that you can use the Deephaven R Client to perform real-time data analysis on streaming data! Of course,
there are performance and memory considerations when pulling data from the server, so it is best to use the provided
TableHandle methods to perform as much of your analysis as possible on the server, and to only pull the data when
something \emph{must} be done in R, like plotting or writing to a local file.
}

\section{Powerful table operations}{

Much of the power of Deephaven's suite of table operations is achieved through the use of the \code{\link[=UpdateBy]{update_by()}}
and \code{\link[=AggBy]{agg_by()}} methods. These table methods are important enough to warrant their own documentation pages, accessible
by clicking on their names, or by running \code{?UpdateBy} or \code{?AggBy}. These methods come with their own suites of functions,
prefixed with \code{agg_} and \code{uby_} respectively, that are discoverable from their documentation pages. Running \code{ls("package:rdeephaven")}
will reveal that most of the functions included in this package are for these methods, so it is important to get acquainted
with them.
}

\section{Getting help}{

While we've done our best to provide good documentation for this package, you may find you need more help than what
this documentation has to offer. If this is the case, please visit the official Deephaven Community Core
\href{https://deephaven.io/core/docs/tutorials/quickstart/}{documentation} to learn more about Deephaven and to find examples
that may help you. If this is not sufficient, feel free to reach out to us on the Deephaven \href{https://deephaven.io/slack}{Community Slack channel}.
We hope you find real-time data analysis in R to be as easy as possible.
}

