% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\name{uby_cum_max}
\alias{uby_cum_max}
\title{Cumulative maximum of specified columns by group}
\arguments{
\item{cols}{String or list of strings denoting the column(s) to operate on. Can be renaming expressions, i.e. “new_col = col”.
Default is to compute the cumulative maximum for all non-grouping columns.}
}
\value{
\code{UpdateByOp} to be used in a call to \code{update_by()}.
}
\description{
Creates a cumulative maximum UpdateByOp that computes the cumulative maximum of each column in \code{cols} for each aggregation group.
}
\details{
This function acts on aggregation groups specified with the \code{by} parameter of the \code{update_by()} caller function.
The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the
\code{A} and \code{B} columns.

This function, like other Deephaven \code{uby} functions, is a generator function. That is, its output is another
function called an \code{UpdateByOp} intended to be used in a call to \code{update_by()}. This detail is typically
hidden from the user by \code{update_by()}, which calls the \code{UpdateByOp} internally. However, it is important to
understand this detail for debugging purposes as the output of a \code{uby} function can otherwise seem unexpected.
}
\examples{
print("hello!")

}
