% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\name{uby_cum_sum}
\alias{uby_cum_sum}
\title{Cumulative sum of specified columns by group}
\arguments{
\item{cols}{String or list of strings denoting the column(s) to operate on. Can be renaming expressions, i.e. “new_col = col”.
Default is to compute the cumulative sum for all non-grouping columns.}
}
\value{
UpdateByOp to be used in \code{update_by()}.
}
\description{
Creates a cumulative sum UpdateByOp that computes the cumulative sum of each column in \code{cols} for each aggregation group.
}
\details{
The aggregation groups that this function acts on are defined with the \code{by} parameter of the \code{update_by()} caller
function. The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the
\code{A} and \code{B} columns.

This function, like the other Deephaven \verb{uby_*} functions, is a generator function. That is, its output is another
function that is intended to be used in a call to \code{update_by()}. This detail is typically hidden from the user by
\code{update_by()}, which calls the generated functions internally. However, it is important to understand this detail
for debugging purposes, as the output of an \verb{agg_*} function can otherwise seem expected.
}
\examples{
print("hello!")

}
