% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\name{uby_delta}
\alias{uby_delta}
\title{Row-wise difference by group}
\arguments{
\item{cols}{String or list of strings denoting the column(s) to operate on. Can be renaming expressions, i.e. “new_col = col”.
Default is to apply the delta operation to all non-grouping columns.}

\item{delta_control}{Defines how the delta operation handles null values. Defaults to \code{'null_dominates'}.}
}
\value{
UpdateByOp to be used in \code{update_by()}.
}
\description{
Creates a delta UpdateByOp for each column in \code{cols}. The Delta operation computes the difference between the
current value and the previous value. When the current value is null, this operation will output null.
When the current value is valid, the output will depend on the \code{delta_control} provided.
}
\details{
The \code{delta_control} parameter controls how this operation treats null values. It can be one of the following:
\itemize{
\item \code{'null_dominates'}: A valid value following a null value returns null.
\item \code{'value_dominates'}: A valid value following a null value returns the valid value.
\item \code{'zero_dominates'}: A valid value following a null value returns zero.
}

The aggregation groups that this function acts on are defined with the \code{by} parameter of the \code{update_by()} caller
function. The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the
\code{A} and \code{B} columns.

This function, like the other Deephaven \verb{uby_*} functions, is a generator function. That is, its output is another
function that is intended to be used in a call to \code{update_by()}. This detail is typically hidden from the user by
\code{update_by()}, which calls the generated functions internally. However, it is important to understand this detail
for debugging purposes, as the output of an \verb{agg_*} function can otherwise seem expected.
}
\examples{
print("hello!")

}
