% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\name{uby_emmin_tick}
\alias{uby_emmin_tick}
\title{Exponential moving minimum by group with ticks as the decay unit}
\arguments{
\item{decay_ticks}{Numeric scalar denoting the decay rate in ticks.}

\item{cols}{String or list of strings denoting the column(s) to operate on. Can be renaming expressions, i.e. “new_col = col”.
Default is to compute the exponential moving minimum for all non-grouping columns.}

\item{operation_control}{OperationControl that defines how special cases will behave. See \code{?op_control} for more information.}
}
\value{
\code{UpdateByOp} to be used in a call to \code{update_by()}.
}
\description{
Creates an exponential moving minimum (EMMIN) UpdateByOp for each column in \code{cols}, using ticks as the decay unit.
}
\details{
The formula used is
\deqn{a = e^{\frac{-1}{\tau}}}
\deqn{\min_0(X) = x_0}
\deqn{\min_i(X) = \min(a*\min_{i-1}(X), \; x_i)}

Where:
\itemize{
\item \eqn{\tau} is \code{decay_ticks}, an input parameter to the method.
\item \eqn{\min_i(X)} is the exponential moving minimum of column \eqn{X} at step \eqn{i}.
\item \eqn{x_i} is the current value.
\item \eqn{i} denotes the time step, ranging from \eqn{i=1} to \eqn{i = n-1}, where \eqn{n} is the number of elements in \eqn{X}.
}

This function acts on aggregation groups specified with the \code{by} parameter of the \code{update_by()} caller function.
The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the
\code{A} and \code{B} columns.

This function, like other Deephaven \code{uby} functions, is a generator function. That is, its output is another
function called an \code{UpdateByOp} intended to be used in a call to \code{update_by()}. This detail is typically
hidden from the user. However, it is important to understand this detail for debugging purposes, as the output of
a \code{uby} function can otherwise seem unexpected.
}
\examples{
print("hello!")

}
