% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\name{uby_ems_time}
\alias{uby_ems_time}
\title{Exponential moving sum by group with time as the decay unit}
\arguments{
\item{decay_time}{ISO-8601-formatted string specifying the decay rate.}

\item{cols}{String or list of strings denoting the column(s) to operate on. Can be renaming expressions, i.e. “new_col = col”.
Default is to compute the exponential moving sum for all non-grouping columns.}

\item{operation_control}{OperationControl that defines how special cases will behave. See \code{?op_control} for more information.}
}
\value{
UpdateByOp to be used in \code{update_by()}.
}
\description{
Creates an exponential moving sum (EMS) UpdateByOp for each column in \code{cols}, using time as the decay unit.
}
\details{
The formula used is
\deqn{a_i = e^{\frac{-dt_i}{\tau}}}
\deqn{\mathcal{S}_0 = x_0}
\deqn{\mathcal{S}_i = a_i*\mathcal{S}_{i-1} + x_i}

Where:
\itemize{
\item \eqn{dt_i} is the difference between time \eqn{t_i} and \eqn{t_{i-1}} in nanoseconds.
\item \eqn{\tau} is \code{decay_time} in nanoseconds, an input parameter to the method.
\item \eqn{\mathcal{S}_i} is the exponential moving sum of column \eqn{X} at time step \eqn{i}.
\item \eqn{x_i} is the current value.
\item \eqn{i} denotes the time step, ranging from \eqn{i=1} to \eqn{i = n-1}, where \eqn{n} is the number of elements in \eqn{X}.
}

The aggregation groups that this function acts on are defined with the \code{by} parameter of the \code{update_by()} caller
function. The aggregation groups are defined by the unique combinations of values in the \code{by} columns. For example,
if \code{by = c("A", "B")}, then the aggregation groups are defined by the unique combinations of values in the
\code{A} and \code{B} columns.

This function, like the other Deephaven \verb{uby_*} functions, is a generator function. That is, its output is another
function that is intended to be used in a call to \code{update_by()}. This detail is typically hidden from the user by
\code{update_by()}, which calls the generated functions internally. However, it is important to understand this detail
for debugging purposes, as the output of a \verb{uby_*} function can otherwise seem unexpected.
}
\examples{
print("hello!")

}
