% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\docType{class}
\name{UpdateBy}
\alias{UpdateBy}
\title{Deephaven's UpdateBy Operations}
\description{
Deephaven's \code{update_by()} table method and suite of \code{uby} functions enable cumulative and moving calculations
on static \emph{and} streaming tables. Complex operations like cumulative minima and maxima, exponential moving averages,
and rolling standard deviations are all possible and effortless to execute. As is always the case in Deephaven,
the results of these calculations will continue to update as their parent tables are updated. Additionally, it's easy
to group data by one or more columns, enabling complex group-wise calculations with a single line of code.
}
\section{Applying UpdateBy operations to a table}{

The table method \code{update_by()} is the entry point for UpdateBy operations. It takes two arguments: the first is an
\code{UpdateByOp} or a list of \code{UpdateByOp}s denoting the calculations to perform on specific columns of the
table. Then, it takes a column name or a list of column names that define the groups to perform the calculations on.
If you don't want grouped calculations, this argument can be omitted.

The \code{update_by()} method itself does not know anything about the columns you want to perform calculations on. Rather,
the desired columns are passed to individual \code{uby} functions, enabling a massive amount of flexibility.
}

\section{\code{uby} functions}{

\code{uby} functions are the workers that actually execute the complex UpdateBy calculations. These functions are
\emph{generators}, meaning they return \emph{functions} that the Deephaven engine knows how to interpret. We call the functions
that they return \code{UpdateByOp}s. These \code{UpdateByOp}s are not R-level functions, but Deephaven-specific data types that
perform all of the intensive calculations. Here is a list of all \code{uby} functions available in Deephaven:
\itemize{
\item \code{\link[=uby_cum_min]{uby_cum_min()}}
\item \code{\link[=uby_cum_max]{uby_cum_max()}}
\item \code{\link[=uby_cum_sum]{uby_cum_sum()}}
\item \code{\link[=uby_cum_prod]{uby_cum_prod()}}
\item \code{\link[=uby_forward_fill]{uby_forward_fill()}}
\item \code{\link[=uby_delta]{uby_delta()}}
\item \code{\link[=uby_emmin_tick]{uby_emmin_tick()}}
\item \code{\link[=uby_emmin_time]{uby_emmin_time()}}
\item \code{\link[=uby_emmax_tick]{uby_emmax_tick()}}
\item \code{\link[=uby_emmax_time]{uby_emmax_time()}}
\item \code{\link[=uby_ems_tick]{uby_ems_tick()}}
\item \code{\link[=uby_ems_time]{uby_ems_time()}}
\item \code{\link[=uby_ema_tick]{uby_ema_tick()}}
\item \code{\link[=uby_ema_time]{uby_ema_time()}}
\item \code{\link[=uby_emstd_tick]{uby_emstd_tick()}}
\item \code{\link[=uby_emstd_time]{uby_emstd_time()}}
\item \code{\link[=uby_rolling_count_tick]{uby_rolling_count_tick()}}
\item \code{\link[=uby_rolling_count_time]{uby_rolling_count_time()}}
\item \code{\link[=uby_rolling_group_tick]{uby_rolling_group_tick()}}
\item \code{\link[=uby_rolling_group_time]{uby_rolling_group_time()}}
\item \code{\link[=uby_rolling_min_tick]{uby_rolling_min_tick()}}
\item \code{\link[=uby_rolling_min_time]{uby_rolling_min_time()}}
\item \code{\link[=uby_rolling_max_tick]{uby_rolling_max_tick()}}
\item \code{\link[=uby_rolling_max_time]{uby_rolling_max_time()}}
\item \code{\link[=uby_rolling_sum_tick]{uby_rolling_sum_tick()}}
\item \code{\link[=uby_rolling_sum_time]{uby_rolling_sum_time()}}
\item \code{\link[=uby_rolling_prod_tick]{uby_rolling_prod_tick()}}
\item \code{\link[=uby_rolling_prod_time]{uby_rolling_prod_time()}}
\item \code{\link[=uby_rolling_avg_tick]{uby_rolling_avg_tick()}}
\item \code{\link[=uby_rolling_avg_time]{uby_rolling_avg_time()}}
\item \code{\link[=uby_rolling_wavg_tick]{uby_rolling_wavg_tick()}}
\item \code{\link[=uby_rolling_wavg_time]{uby_rolling_wavg_time()}}
\item \code{\link[=uby_rolling_std_tick]{uby_rolling_std_tick()}}
\item \code{\link[=uby_rolling_std_time]{uby_rolling_std_time()}}
}

For more details on each aggregation function, click on one of the methods above or see the reference documentation
by running \code{?uby_cum_min}, \code{?uby_delta}, etc.
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{NULL
}\if{html}{\out{</div>}}
}

