% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_ops_wrapper.R
\docType{class}
\name{UpdateByOp}
\alias{UpdateByOp}
\title{Deephaven UpdateByOps}
\description{
An \code{UpdateByOp} is the return type of one of Deephaven's \code{\link[=UpdateBy]{uby}} functions. It is a function that performs
the computation specified by the \code{uby} function. These are intended to be passed directly to \code{update_by()},
and should never be instantiated directly be user code. For more information, see the vignette on
\code{uby} functions with \code{vignette("update_by")}.

If multiple tables have the same schema and the same UpdateBy operations need to be applied to each table, saving
these objects directly in a variable may be useful to avoid having to re-create them each time:

\if{html}{\out{<div class="sourceCode">}}\preformatted{operations <- c(uby_rolling_avg_tick("XAvg = X", "YAvg = Y"),
                uby_rolling_std_tick("XStd = X", "YStd = Y"))

result1 <- th1$update_by(operations, by="Group")
result2 <- th2$update_by(operations, by="Group")
}\if{html}{\out{</div>}}

In this example, \code{operations} would be a vector of two \code{UpdateByOp}s that can be reused in multiple calls to \code{update_by()}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdateByOp-new}{\code{UpdateByOp$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdateByOp-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdateByOp-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdateByOp$new(update_by_op)}\if{html}{\out{</div>}}
}

}
}
