% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_wrapper.R
\docType{class}
\name{Client}
\alias{Client}
\title{The Deephaven Client}
\description{
The Deephaven Client class is responsible for establishing and maintaining
a connection to a running Deephaven server and facilitating basic server requests.
}
\examples{

# connect to the Deephaven server running on "localhost:10000" using anonymous 'default' authentication
client_options <- ClientOptions$new()
client <- Client$new(target="localhost:10000", client_options=client_options)

# open a table that already exists on the server
new_table_handle1 <- client$open_table("table_on_the_server")

# create a new dataframe, import onto the server, and retrieve a reference
new_data_frame <- data.frame(matrix(rnorm(10 * 1000), nrow = 10))
new_table_handle2 <- client$import_table(new_data_frame)

# run a python script on the server (default client options specify a Python console)
client$run_script("print([i for i in range(10)])")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Client-new}{\code{Client$new()}}
\item \href{#method-Client-open_table}{\code{Client$open_table()}}
\item \href{#method-Client-import_table}{\code{Client$import_table()}}
\item \href{#method-Client-run_script}{\code{Client$run_script()}}
\item \href{#method-Client-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-new"></a>}}
\if{latex}{\out{\hypertarget{method-Client-new}{}}}
\subsection{Method \code{new()}}{
Connect to a running Deephaven server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(target, client_options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{The address of the Deephaven server.}

\item{\code{client_options}}{ClientOptions instance with the parameters needed to connect to the server.
See ?ClientOptions for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-open_table"></a>}}
\if{latex}{\out{\hypertarget{method-Client-open_table}{}}}
\subsection{Method \code{open_table()}}{
Opens a table named 'name' from the server if it exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$open_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the table to open from the server, passed as a string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TableHandle reference to the requested table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-import_table"></a>}}
\if{latex}{\out{\hypertarget{method-Client-import_table}{}}}
\subsection{Method \code{import_table()}}{
Imports a new table to the Deephaven server. Note that this new table is not automatically bound to
a variable name on the server. See `?TableHandle` for more information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$import_table(table_object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_object}}{An R Data Frame, a dplyr Tibble, an Arrow Table, or an Arrow RecordBatchReader
containing the data to import to the server.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TableHandle reference to the new table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-run_script"></a>}}
\if{latex}{\out{\hypertarget{method-Client-run_script}{}}}
\subsection{Method \code{run_script()}}{
Runs a script on the server. The script must be in the language that the server console was started with.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$run_script(script)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{script}}{Code to be executed on the server, passed as a string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Client-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
