% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_options_wrapper.R
\name{ClientOptions}
\alias{ClientOptions}
\title{Deephaven ClientOptions}
\description{
Client options provide a simple interface to the Deephaven server's authentication protocols.
This makes it easy to connect to a Deephaven server with any flavor of authentication, and shields the API from
any future changes to the underlying implementation.

Currently, we support three different kinds of authentication that your Deephaven server might be using:

- "default": Default (or anonymous) authentication does not require any username or password. If you are
   running the Deephaven server locally, this is probably the kind of authentication you are using.

- "basic": Basic authentication requires a standard username and password pair.

- "custom": Custom authentication requires general key-value pairs.

In addition to setting your authentication parameters when you connect to a client, you can also start
a console in one of our supported server languages. We currently support Python and Groovy, and if you
want start a console upon client connection, you must ensure that the server you're connecting to was
started with support for the language you want to use.
}
\examples{

# connect to the Deephaven server running on "localhost:10000" with anonymous 'default' authentication
client_options <- ClientOptions$new()
client_options$set_default_authentication()
client <- Client$new(target="localhost:10000", client_options=client_options)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientOptions-new}{\code{ClientOptions$new()}}
\item \href{#method-ClientOptions-set_default_authentication}{\code{ClientOptions$set_default_authentication()}}
\item \href{#method-ClientOptions-set_basic_authentication}{\code{ClientOptions$set_basic_authentication()}}
\item \href{#method-ClientOptions-set_custom_authentication}{\code{ClientOptions$set_custom_authentication()}}
\item \href{#method-ClientOptions-set_session_type}{\code{ClientOptions$set_session_type()}}
\item \href{#method-ClientOptions-clone}{\code{ClientOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-set_default_authentication"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-set_default_authentication}{}}}
\subsection{Method \code{set_default_authentication()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$set_default_authentication()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-set_basic_authentication"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-set_basic_authentication}{}}}
\subsection{Method \code{set_basic_authentication()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$set_basic_authentication(username, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-set_custom_authentication"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-set_custom_authentication}{}}}
\subsection{Method \code{set_custom_authentication()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$set_custom_authentication(auth_key, auth_value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-set_session_type"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-set_session_type}{}}}
\subsection{Method \code{set_session_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$set_session_type(session_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
